backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
fairlearn==0.10.0
shap==0.45.1
reportlab==4.1.0
python-multipart==0.0.9
joblib==1.4.2
scipy==1.13.0
google-generativeai==0.7.2
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.0
